Move zk mock classes to dedicated folder
