+++ examples/bpi-monolith-test-suite/bpi/components/zkmock/agreementalt.ts
