+++ examples/bpi-monolith-test-suite/bpi/components/zkmock/predicates.ts
