+++ examples/bpi-monolith-test-suite/bpi/components/zkmock/shield.ts
