+++ examples/bpi-monolith-test-suite/bpi/components/zkmock/verifier.ts
